Continue 64 bits

+++ SparkDaemon/Sources/SDAEHandlers.c
@@ -23,7 +23,7 @@ OSStatus SDGetEditorIsTrapping(Boolean *trapping) {
   /* Check front process */
   ProcessInfoRec info;
   bzero(&info, sizeof(info));
-  info.processInfoLength = sizeof(info);
+  info.processInfoLength = (UInt32)sizeof(info);
   err = GetProcessInformation(&psn, &info);
   require_noerr(err, bail);
